moved register mixin into register view.
