+++ huxley/api/mixins.py
@@ -7,46 +7,7 @@ from django.db import transaction
 from rest_framework import status
 from rest_framework.response import Response
 
-from huxley.core.models import Delegate, School
-
-
-class RegisterMixin(object):
-    """
-    Mixin customized to create necessary model instances when registering
-    """
-
-    def create(self, request, *args, **kwargs):
-        user_data = request.data['user']
-        registration_data = request.data['registration']
-        user_serializer = self.serializer_classes['user'](data=user_data)
-        registration_serializer = self.serializer_classes['registration'](
-            data=registration_data)
-        is_user_valid = user_serializer.is_valid()
-        is_registration_valid = self.is_registration_valid(
-            registration_serializer)
-
-        if is_user_valid and is_registration_valid:
-            user_serializer.save()
-            school_id = School.objects.get(name=user_data['school']['name']).id
-            registration_data['school'] = school_id
-            registration_serializer = self.serializer_classes['registration'](
-                data=registration_data)
-            registration_serializer.is_valid()
-            registration_serializer.save()
-            data = {'user': user_serializer.data,
-                    'registration': registration_serializer.data}
-            response_status = status.HTTP_200_OK
-        else:
-            data = registration_serializer.errors.copy()
-            data.update(user_serializer.errors)
-            response_status = status.HTTP_400_BAD_REQUEST
-
-        return Response(data, status=response_status)
-
-    def is_registration_valid(self, registration_serializer):
-        registration_serializer.is_valid()
-        errors = registration_serializer.errors
-        return len(errors) == 1 and 'school' in errors
+from huxley.core.models import Delegate
 
 
 class ListUpdateModelMixin(object):
